MINOR: allowing for a non-array value

+++ code/api/EcommerceCache.php
@@ -8,7 +8,6 @@ class EcommerceCache extends Object implements flushable
 {
 
     private static $cache_in_mysql_tables = array(
-        'ProductGroup'
     );
 
     public static function flush()
@@ -16,44 +15,48 @@ class EcommerceCache extends Object implements flushable
         $cache = SS_Cache::factory('any');
         $cache->clean(Zend_Cache::CLEANING_MODE_ALL);
         $tables = Config::inst()->get('EcommerceCache', 'cache_in_mysql_tables');
-        foreach($tables as $table) {
-            $table = self::make_mysql_table_name($table);
-            DB::query(
-                '
-                DROP TABLE IF EXISTS "'.$table.'";
-                '
-            );
-            DB::query(
-                '
-                CREATE TABLE "'.$table.'" (
-                  "PAGEID" int(11) NOT NULL,
-                  "CACHEKEY" CHAR(50) NOT NULL,
-                  "DATA" TEXT
-                ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
-                '
-            );
+        if(is_array($tables)) {
+            foreach($tables as $table) {
+                $table = self::make_mysql_table_name($table);
+                DB::query(
+                    '
+                    DROP TABLE IF EXISTS "'.$table.'";
+                    '
+                );
+                DB::query(
+                    '
+                    CREATE TABLE "'.$table.'" (
+                      "PAGEID" int(11) NOT NULL,
+                      "CACHEKEY" CHAR(50) NOT NULL,
+                      "DATA" TEXT
+                    ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+                    '
+                );
 
-            DB::query(
-                '
-                ALTER TABLE "'.$table.'"
-                  ADD PRIMARY KEY ("PAGEID","CACHEKEY"),
-                  ADD KEY "CACHEKEY" ("CACHEKEY"),
-                  ADD KEY "PAGEID" ("PAGEID");
-                '
-            );
+                DB::query(
+                    '
+                    ALTER TABLE "'.$table.'"
+                      ADD PRIMARY KEY ("PAGEID","CACHEKEY"),
+                      ADD KEY "CACHEKEY" ("CACHEKEY"),
+                      ADD KEY "PAGEID" ("PAGEID");
+                    '
+                );
+            }
         }
     }
 
     private static function clean()
     {
         $tables = Config::inst()->get('EcommerceCache', 'cache_in_mysql_tables');
-        foreach($tables as $table) {
-            $table = self::make_mysql_table_name($table);
-            DB::query(
-                '
-                DELETE FROM "'.$table.'";
-                '
-            );
+        if(is_array($tables)) {
+            foreach($tables as $table) {
+                $table = self::make_mysql_table_name($table);
+                DB::query(
+                    '
+                    DELETE FROM "'.$table.'";
+                    '
+                );
+            }
         }
     }
 
@@ -68,7 +71,7 @@ class EcommerceCache extends Object implements flushable
     public static function load($table, $id, $cacheKey)
     {
         $tables = Config::inst()->get('EcommerceCache', 'cache_in_mysql_tables');
-        if(in_array($table, $tables)) {
+        if(is_array($tables) && in_array($table, $tables)) {
             $table = self::make_mysql_table_name($table);
             $id = (int)$id;
             if(isset(self::$_items[$table])) {
@@ -119,7 +122,7 @@ class EcommerceCache extends Object implements flushable
     public static function save($table, $id, $cacheKey, $data)
     {
         $tables = Config::inst()->get('EcommerceCache', 'cache_in_mysql_tables');
-        if(in_array($table, $tables)) {
+        if(is_array($tables) && in_array($table, $tables)) {
             $table = self::make_mysql_table_name($table);
             $id = (int)$id;
             if(strlen($cacheKey) > 50) {
